+++ doc/source/changelog.md
@@ -17,8 +17,9 @@ Primarily developped with Godot 4.2.
 - Saving with `save_all_modified_blocks` now automatically flushes eventual caches implemented by `VoxelStream` upon completion
 - Added `VoxelStreamMemory`, which stores in memory instead of the filesystem. This is mainly for testing purposes.
 - More memory allocations are now tracked by Godot (you might notice `OS.get_static_memory_usage()` returns slightly more)
-- `VoxelGeneratorGraph`: Added GPU support for the `Select` node
 - `VoxelBuffer`: exposed `fill_area_f`
+- `VoxelEngine`: added methods to get the version of the voxel engine
+- `VoxelGeneratorGraph`: Added GPU support for the `Select` node
 - `VoxelTool`:
     - Added `grow_sphere` as alternate way to progressively grow or shrink matter in a spherical region with smooth voxels (thanks to Piratux)
 - `VoxelToolBuffer`: edits are now allowed even if the affected area is partially out of bounds of the target buffer. Results will be clipped.